+++ index.html
@@ -10,6 +10,7 @@
     <div id="app">
       <section class="item-view"></section>
     </div>
+    <div id="movie-detail"></div>
     <div id="toast_message"></div>
     <script src="./src/index.js"></script>
   </body>
